+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 What's changed in UniteJS VSCode extension.
 
+## v1.0.3
+
+* Added CSS Linter support
+* Added platform-cordova-dev and platform-cordova-theme commands
+* Added e2e-install options for drivers
+* Fixed e2e single command
+
 ## v1.0.2
 
 * Added PWA buildConfiguration option